Question 4

+++ TP/TP3/main.c
@@ -13,7 +13,7 @@ int main(int argc, const char * argv[]) {
     printf("cent = %d", cent);
 }
 
-long int intervalle_ms(Time debut, Time fin) {
+int intervalle_ms(Time debut, Time fin) {
     return (fin.tv_sec - debut.tv_sec) * 1000 + (fin.tv_usec - debut.tv_usec) / 1000;
 }
